+++ src/js/chameleon-no-jquery.js
@@ -1,8 +1,0 @@
-/**
- * For websites that already provides jQuery.
- *
- * Examples: MediaWiki (en.opensuse.org), WordPress (news.opensuse.org)
- */
-
-require("bootstrap");
-require("./chameleon-no-bootstrap");