+++ src/js/chameleon.js
@@ -1,2 +1,9 @@
-require("jquery");
-require("./chameleon-no-jquery");
+/**
+ * For webistes that already provide Bootstrap JavaScript.
+ *
+ * Examples: Weblate (l10n.opensuse.org), Cachet (status.opensuse.org)
+ */
+
+require("./components/megamenu");
+require("./components/toc");
+require("./components/icons");