+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+
+- js: `jquery.js` and `bootstrap.js` are not bundled anymore **BREAKING CHANGE**
+- js,icon: optimize size
+
 ## [2.0.0] - 2020-03-04
 
 ### Added